Clean up least recently used entries in external resources file store
The files in the external resources file store cache are now
automatically cleaned up based on a least recently used strategy. The
structure in the cache is as follows:

 - modules-2
   - resources-2.1
     - 0
       - hash 1 (*)
       - hash 2 (*)
       - ...
     - ...

The cache now keeps track of files accessed inside the marked hash
directories. The cleanup action then deletes all such files that have
not been accessed in the last 30 days.