Clean up least recently used entries in module metadata store
The files in the module metadata store cache are now automatically
cleaned up based on a least recently used strategy. The structure in
the cache is as follows:

 - modules-2
   - metadata-2.58
     - descriptors
       - groupId
         - artifactId
           - version
             - hash1 (*)
               - file
             - hash2 (*)
               - file

The cache now keeps track of files accessed inside the marked hash
directories. The cleanup action then deletes all such files that have
not been accessed in the last 30 days.