Bump major cache version of root modules cache
Incrementing the major version avoids having to delete the complete
`modules-2` cache because it has to be reinitialized. This commit
ignores the existing cross-version tests when the major version changes
because in that case no cached artifacts can be reused. It adds an
additional cross-version test that checks that all artifacts need to
be downloaded again.